+ store config for prod
